# openapi/shipflow-api.yaml
openapi: 3.0.3
info:
  title: ShipFlow API
  version: 1.0.0
  description: |
    ShipFlow 跨境物流履约与费用结算系统第一版 API 契约。
    所有租户业务接口从 JWT 身份上下文获取 tenant_id，不接受调用方任意传入。
servers:
  - url: /api/v1
    description: ShipFlow API v1
...
security:
  - bearerAuth: []
paths:
...
  /exceptions:
    get:
      tags: [Exceptions]
      summary: 查询当前租户异常单
      operationId: listExceptionCases
      parameters: [{ name: orderId, in: query, schema: { $ref: '#/components/schemas/Id' } }, { name: status, in: query, schema: { type: string, enum: [OPEN, PROCESSING, WAITING_PROVIDER_FEEDBACK, RESOLVED, PENDING_FINANCE_CONFIRMATION, CLOSED] } }, { name: workbenchFilter, in: query, description: 仅工作台使用；按调用方店铺范围过滤。, schema: { type: string, enum: [PENDING_FOLLOW_UP] } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }]
      responses: { '200': { description: 异常单分页, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessExceptionCasePage' } } } }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /exceptions/{exceptionId}:
    get:
      tags: [Exceptions]
      summary: 查询当前租户异常单详情
      operationId: getExceptionCase
      parameters: [{ name: exceptionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }]
      responses: { '200': { description: 异常单详情及索赔摘要, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessExceptionCase' } } } }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /orders/{orderId}/exceptions:
    post:
      tags: [Exceptions]
      summary: 从订单或其轨迹事件创建异常单
      operationId: createExceptionCase
      parameters: [{ $ref: '#/components/parameters/OrderId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateExceptionRequest' } } } }
      responses: { '201': { description: 创建成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessExceptionCase' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ExceptionClaimError' } }
  /exceptions/{exceptionId}/assign:
    post:
      tags: [Exceptions]
      summary: 分派异常单并从 OPEN 进入 PROCESSING
      operationId: assignExceptionCase
      parameters: [{ name: exceptionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AssignExceptionRequest' } } } }
      responses: { '200': { description: 分派成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessExceptionCase' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
  /exceptions/{exceptionId}/status:
    post:
      tags: [Exceptions]
      summary: 解决或关闭异常单
      operationId: transitionExceptionCase
      description: 仅允许按异常状态机前向推进；数据库时间按 UTC 存储，接口时间为 UTC ISO-8601，前端展示转换为 Asia/Shanghai。
      parameters: [{ name: exceptionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ExceptionStatusRequest' } } } }
      responses: { '200': { description: 状态推进成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessExceptionCase' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
  /exceptions/{exceptionId}/handling-records:
    get:
      tags: [Exceptions]
      summary: 查询异常处理记录
      operationId: listExceptionHandlingRecords
      parameters: [{ name: exceptionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }]
      responses: { '200': { description: 处理记录列表, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ExceptionHandlingRecordView' } } } } }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
    post:
      tags: [Exceptions]
      summary: 追加异常处理记录
      operationId: createExceptionHandlingRecord
      parameters: [{ name: exceptionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateHandlingRecordRequest' } } } }
      responses: { '201': { description: 记录创建成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessExceptionHandlingRecord' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
  /exceptions/{exceptionId}/evidence:
    get:
      tags: [Exceptions]
      summary: 查询异常证据附件元数据
      operationId: listExceptionEvidence
      parameters: [{ name: exceptionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }]
      responses: { '200': { description: 证据附件元数据列表, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ExceptionEvidenceView' } } } } }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
    post:
      tags: [Exceptions]
      summary: 上传异常证据附件
      operationId: uploadExceptionEvidence
      parameters: [{ name: exceptionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { multipart/form-data: { schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } } } } }
      responses: { '201': { description: 上传成功或内容哈希去重, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessExceptionEvidence' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
  /exceptions/{exceptionId}/evidence/{attachmentId}/content:
    get:
      tags: [Exceptions]
      summary: 下载异常证据附件内容
      operationId: downloadExceptionEvidence
      parameters: [{ name: exceptionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }, { name: attachmentId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }]
      responses: { '200': { description: 附件二进制内容, content: { application/octet-stream: { schema: { type: string, format: binary } } } }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /exceptions/{exceptionId}/claim:
    post:
      tags: [Exceptions]
      summary: 为已解决且符合条件的异常单创建索赔
      operationId: createClaim
      parameters: [{ name: exceptionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateClaimRequest' } } } }
      responses: { '201': { description: 创建成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessClaim' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ExceptionClaimError' } }
  /claims/{claimId}:
    get:
      tags: [Exceptions]
      summary: 查询当前租户索赔详情
      operationId: getClaim
      parameters: [{ name: claimId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }]
      responses: { '200': { description: 索赔详情, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessClaim' } } } }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /claims/{claimId}/submit:
    post:
      tags: [Exceptions]
      summary: 提交索赔
      operationId: submitClaim
      parameters: [{ name: claimId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClaimActionRequest' } } } }
      responses: { '200': { description: 提交成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessClaim' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
  /claims/{claimId}/result:
    post:
      tags: [Exceptions]
      summary: 审核通过或驳回索赔
      operationId: resolveClaim
      parameters: [{ name: claimId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClaimResultRequest' } } } }
      responses: { '200': { description: 审核完成, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessClaim' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
...
  /claims/{claimId}/close:
    post:
      tags: [Exceptions]
      summary: 关闭已审核索赔
      operationId: closeClaim
      parameters: [{ name: claimId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClaimActionRequest' } } } }
      responses: { '200': { description: 关闭成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiSuccessClaim' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/AuthError' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: REFRESH_TOKEN
      description: HttpOnly Refresh Token Cookie；客户端不可读取。
    providerHmac:
      type: apiKey
      in: header
      name: X-Provider-Signature
      description: Base64 编码的 HMAC-SHA256 签名，签名原文为 timestamp + "." + rawRequestBody。
  headers:
    TraceId:
      description: 与响应体 traceId 一致的链路追踪 ID。
      schema:
        type: string
    NoStore:
      description: 禁止缓存认证凭证相关响应。
      schema:
        type: string
        enum: [no-store]
        default: no-store
    WWWAuthenticate:
      description: RFC 6750 Bearer 认证挑战；具体错误详情不在 Header 中泄露敏感信息。
      schema:
        type: string
        example: Bearer
...
  parameters:
...
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: 请求参数或格式错误。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    AuthError:
      description: 未登录、Token 无效或已过期。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Forbidden:
      description: 权限不足或跨租户访问。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: 资源不存在。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 幂等、乐观锁或订单状态冲突。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    ExceptionClaimError:
      description: 异常或索赔业务前置条件不满足。
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    Id:
      type: string
      pattern: '^[0-9]+$'
      example: '10001'
...
    UtcDateTime:
      type: string
      format: date-time
      example: '2026-08-02T10:30:00.000Z'
      description: UTC ISO-8601 时间。
...
    ApiSuccess:
      type: object
      required: [success, traceId, data]
      properties:
        success:
          type: boolean
          enum: [true]
        traceId:
          type: string
        message:
          type: string
          example: OK
        data: {}
...
    ApiError:
      type: object
      required: [success, traceId, error]
      properties:
        success:
          type: boolean
          enum: [false]
        traceId:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: ORDER-1001
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    PageMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    CreateExceptionRequest:
      type: object
      required: [exceptionType, description, reportedAt]
      properties:
        exceptionType: { type: string, enum: [ADDRESS, CUSTOMS, TRANSPORT, OTHER] }
        description: { type: string, minLength: 1, maxLength: 1000 }
        reportedAt: { $ref: '#/components/schemas/UtcDateTime' }
        trackingEventId: { $ref: '#/components/schemas/Id', nullable: true, description: 必须属于路径订单及当前租户。 }
    AssignExceptionRequest:
      type: object
      required: [assignedToUserId, responsibleParty, version]
      properties:
        assignedToUserId: { $ref: '#/components/schemas/Id' }
        responsibleParty: { type: string, enum: [MERCHANT, PROVIDER, CUSTOMS, CUSTOMER, OTHER] }
        reason: { type: string, maxLength: 1000 }
        version: { type: integer, format: int64, minimum: 0 }
    ExceptionStatusRequest:
      type: object
      required: [status, reason, version]
      properties:
        status: { type: string, enum: [PROCESSING, WAITING_PROVIDER_FEEDBACK, RESOLVED, PENDING_FINANCE_CONFIRMATION, CLOSED] }
        reason: { type: string, minLength: 1, maxLength: 1000 }
        version: { type: integer, format: int64, minimum: 0 }
    CreateHandlingRecordRequest:
      type: object
      required: [recordType, content]
      properties:
        recordType: { type: string, enum: [CONTACT, FOLLOW_UP, PROVIDER_FEEDBACK, INTERNAL_NOTE, OTHER] }
        content: { type: string, minLength: 1, maxLength: 4000 }
    ExceptionHandlingRecordView:
      type: object
      required: [id, exceptionId, recordNo, handledByUserId, recordType, content, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        exceptionId: { $ref: '#/components/schemas/Id' }
        recordNo: { type: string }
        handledByUserId: { $ref: '#/components/schemas/Id' }
        recordType: { type: string, enum: [CONTACT, FOLLOW_UP, PROVIDER_FEEDBACK, INTERNAL_NOTE, OTHER] }
        content: { type: string }
        createdAt: { $ref: '#/components/schemas/UtcDateTime' }
    ExceptionEvidenceView:
      type: object
      required: [id, exceptionId, uploadedByUserId, originalFileName, contentType, fileSize, contentSha256, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        exceptionId: { $ref: '#/components/schemas/Id' }
        uploadedByUserId: { $ref: '#/components/schemas/Id' }
        originalFileName: { type: string }
        contentType: { type: string }
        fileSize: { type: integer, format: int64, minimum: 1 }
        contentSha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        createdAt: { $ref: '#/components/schemas/UtcDateTime' }
    ApiSuccessExceptionHandlingRecord:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/ExceptionHandlingRecordView' } } }]
    ApiSuccessExceptionEvidence:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/ExceptionEvidenceView' } } }]
    CreateClaimRequest:
      type: object
      required: [claimAmount, currency]
      properties:
        claimAmount: { type: number, format: decimal, minimum: 0.01, multipleOf: 0.01 }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
    ClaimActionRequest:
      type: object
      required: [version]
      properties:
        reason: { type: string, maxLength: 1000 }
        version: { type: integer, format: int64, minimum: 0 }
    ClaimResultRequest:
      type: object
      required: [status, reason, resolvedAt, version]
      properties:
        status: { type: string, enum: [APPROVED, REJECTED] }
        reason: { type: string, minLength: 1, maxLength: 1000 }
        resolvedAt: { $ref: '#/components/schemas/UtcDateTime' }
        version: { type: integer, format: int64, minimum: 0 }
    ClaimView:
      type: object
      required: [id, exceptionId, claimNo, status, claimAmount, currency, version, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        exceptionId: { $ref: '#/components/schemas/Id' }
        claimNo: { type: string }
        status: { type: string, enum: [OPEN, SUBMITTED, APPROVED, REJECTED, CLOSED] }
        claimAmount: { type: number, format: decimal }
        currency: { type: string }
        submittedAt: { $ref: '#/components/schemas/UtcDateTime', nullable: true }
        resolvedAt: { $ref: '#/components/schemas/UtcDateTime', nullable: true }
        version: { type: integer, format: int64 }
        createdAt: { $ref: '#/components/schemas/UtcDateTime' }
        updatedAt: { $ref: '#/components/schemas/UtcDateTime' }
    ExceptionCaseView:
      type: object
      required: [id, orderId, exceptionNo, exceptionType, status, description, reportedAt, version, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        orderId: { $ref: '#/components/schemas/Id' }
        exceptionNo: { type: string }
        exceptionType: { type: string, enum: [ADDRESS, CUSTOMS, TRANSPORT, OTHER] }
        status: { type: string, enum: [OPEN, PROCESSING, WAITING_PROVIDER_FEEDBACK, RESOLVED, PENDING_FINANCE_CONFIRMATION, CLOSED] }
        description: { type: string }
        reportedAt: { $ref: '#/components/schemas/UtcDateTime' }
        assignedToUserId: { $ref: '#/components/schemas/Id', nullable: true }
        storeId: { $ref: '#/components/schemas/Id' }
        responsibleParty: { type: string, enum: [MERCHANT, PROVIDER, CUSTOMS, CUSTOMER, OTHER], nullable: true }
        version: { type: integer, format: int64 }
        createdAt: { $ref: '#/components/schemas/UtcDateTime' }
        updatedAt: { $ref: '#/components/schemas/UtcDateTime' }
        claim: { $ref: '#/components/schemas/ClaimView', nullable: true }
    ExceptionCasePage:
      allOf: [{ $ref: '#/components/schemas/PageMeta' }, { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/ExceptionCaseView' } } } }]
    ApiSuccessExceptionCase:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/ExceptionCaseView' } } }]
    ApiSuccessExceptionCasePage:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/ExceptionCasePage' } } }]
    ApiSuccessClaim:
      allOf: [{ $ref: '#/components/schemas/ApiSuccess' }, { type: object, properties: { data: { $ref: '#/components/schemas/ClaimView' } } }]